backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests>=2.28.0
python-dotenv==0.21.0
pytest==7.4.0
httpx==0.24.1
